requirements.txt
paho-mqtt~=1.5.1
prometheus_client